Level 2-10 : ArithmeticCalculator 클래스에 추가로 나머지 연산(%) 기능을 추가하기 위해 ModOperator 클래스를 만들어 추가합니다.

+++ src/main/java/calculator/CircleCalculator.java
@@ -5,7 +5,7 @@ public class CircleCalculator extends Calculator{
 
     // 사용하지 않는 오버라이딩 메서드를 접근제어자를 protected를 써서 접근 불가능하게 만듬
     @Override
-    protected double calculate(int firstNumber, int secondNumber, char operator) throws InputErrorException {
+    protected int calculate(int firstNumber, int secondNumber, char operator) throws InputErrorException {
         return 0;
     }
